refactor: centralize UsageMetrics type in src/types and fix docs snippet range
Re-export UsageMetrics from src/types/index.ts and update all imports
to use the centralized path. Fix docs:check snippet range after line
number shift from import consolidation.
